+++ djangopypi2/urls.py
@@ -18,6 +18,7 @@ def static_urls():
 
 urlpatterns = patterns('',
     url(r'', include('djangopypi2.apps.pypi_ui.urls')),
+    url(r'', include('djangopypi2.apps.pypi_users.urls')),
     url(r'', include('djangopypi2.apps.pypi_packages.urls')),
     url(r'', include('djangopypi2.apps.pypi_frontend.urls')),
 ) + static_urls()